Coin detail page added

+++ app/src/main/java/com/example/cyrptocurrencyapp/data/model/CoinDataModel.kt
@@ -5,7 +5,7 @@ data class CoinDataModel(
     val symbol: String,
     val name: String,
     val image: String,
-    val market_cap : Long,
+    val market_cap : Double,
     val price: Double,
     val price_percentage_change: Double,
     val low_price: Double,
